+++ Docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 
 ### Changed
 - Updated NuGet packages
+- Increased the limit on the number of download streams for a user with a free license
 
 ## [0.6.550] - 2025-06-29
 ### Fixed